+++ projects/pastanaga-angular/src/lib/modal/modal-dialog/modal-dialog.component.ts
@@ -4,9 +4,6 @@ import {
     ChangeDetectorRef,
     Component,
     ElementRef,
-    Inject,
-    NgZone,
-    OnDestroy,
     ViewChild,
     ViewEncapsulation,
 } from '@angular/core';
@@ -15,8 +12,6 @@ import { ModalRef } from '../modal.model';
 import { TRANSITION_DURATION } from '../../common';
 import { BreakpointObserver } from '../../breakpoint-observer/breakpoint.observer';
 import { takeUntil } from 'rxjs/operators';
-import { DOCUMENT } from '@angular/common';
-import { BehaviorSubject } from 'rxjs';
 
 @Component({
     selector: 'pa-modal-dialog',
@@ -25,8 +20,7 @@ import { BehaviorSubject } from 'rxjs';
     changeDetection: ChangeDetectionStrategy.OnPush,
     encapsulation: ViewEncapsulation.None,
 })
-export class ModalDialogComponent extends BaseModalComponent implements AfterViewInit, OnDestroy {
-    @ViewChild('dialogContainer', { read: ElementRef }) dialogContainer?: ElementRef;
+export class ModalDialogComponent extends BaseModalComponent implements AfterViewInit {
     @ViewChild('header', { read: ElementRef }) header?: ElementRef;
     @ViewChild('image', { read: ElementRef }) image?: ElementRef;
     @ViewChild('description', { read: ElementRef }) description?: ElementRef;
@@ -37,20 +31,12 @@ export class ModalDialogComponent extends BaseModalComponent implements AfterVie
     hasFooter = false;
 
     private _headerHeightSettingDelay = TRANSITION_DURATION.moderate;
-    dialogHeightHasChanged$ = new BehaviorSubject<number>(0);
-    observer: ResizeObserver = new ResizeObserver((entries) => {
-        this.ngZone.runOutsideAngular(() => {
-            this.dialogHeightHasChanged$.next(entries[0].contentRect.top);
-        });
-    });
 
     constructor(
         public ref: ModalRef,
         protected cdr: ChangeDetectorRef,
         private element: ElementRef,
         private breakpoint: BreakpointObserver,
-        private ngZone: NgZone,
-        @Inject(DOCUMENT) private document: any,
     ) {
         super(ref, cdr);
         this.breakpoint.currentMode
@@ -78,16 +64,5 @@ export class ModalDialogComponent extends BaseModalComponent implements AfterVie
                 `${Math.ceil(this.header?.nativeElement.getBoundingClientRect().height)}px`,
             );
         }, this._headerHeightSettingDelay);
-
-        // updating dialog top offset on dialog height change
-        this.observer.observe(this.dialogContainer?.nativeElement);
-        this.dialogHeightHasChanged$.pipe(takeUntil(this._terminator)).subscribe(() => {
-            const dialogRect = this.dialogContainer?.nativeElement.getBoundingClientRect();
-            this.document.documentElement.style.setProperty('--containerTranslateY', `${Math.ceil(dialogRect.top)}`);
-        });
-    }
-
-    ngOnDestroy() {
-        this.observer.unobserve(this.dialogContainer?.nativeElement);
     }
 }