+++ projects/pastanaga-angular/src/lib/popup/popup.component.ts
@@ -118,7 +118,7 @@ export class PopupComponent implements OnInit, OnDestroy {
         if (!this.keepOthersOpen) {
             this.popupService.closeAllButId.next(this.id);
         }
-        this.setPopupPosition(style);
+        this.style = style;
         this.isDisplayed = true;
         this.onOpen.emit();
         if (!this.stayVisible) {
@@ -128,11 +128,6 @@ export class PopupComponent implements OnInit, OnDestroy {
 
         markForCheck(this.cdr);
         this.adjustPosition();
-    }
-
-    setPopupPosition(position: PositionStyle | undefined) {
-        this.style = position;
-        markForCheck(this.cdr);
     }
 
     private adjustPosition() {